requirements.txt
google-adk>=1.25.1
requests>=2.31.0
beautifulsoup4>=4.12.3
scikit-learn>=1.5.0
pypdf>=4.2.0
chromadb>=0.5.23
langchain-core>=0.3.40
langchain-chroma>=0.2.0
